src/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.4.1
Pillow==9.0.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
requests==2.27.1
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.3
